backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
pillow==9.5.0
numpy==1.24.3
opencv-python==4.7.0.72
